Set tilesType to moc for ZTF
I removed the `preComputed_ZTF_pixel_indices_256.dat` file from
the repository because it is too large to effictively track with
git and because we should be evaulating the HEALPix field
footprints from the same vector representations that we use for
the D3 visualization.

@mcoughlin suggests that setting `params["tilesType"] = "moc"`
should work as a temporary workaround.