Speed up unit tests by calling create_db only once
Call `growth.too.models.create_all()` once at pytest session scope,
and then roll back to a clean database at the end of every test.
This speeds up the unit tests significantly because it takes about
40 seconds to run `create_all()`, a cost that was being incurred
once for each unit test.

The rollback is achieved with the help of the
`pytest-flask-sqlalchemy` package.